+++ src/main/java/com/github/theredbrain/foodoverhaul/component/type/FoodBlockDataComponent.java
@@ -7,7 +7,7 @@ import net.minecraft.network.codec.StreamCodec;
 
 public record FoodBlockDataComponent(FoodBlockEntity.FoodBlockData food_block_data) {
 	public static final Codec<FoodBlockDataComponent> CODEC = FoodBlockEntity.FoodBlockData.CODEC.xmap(FoodBlockDataComponent::new, FoodBlockDataComponent::food_block_data);
-	public static final StreamCodec<RegistryFriendlyByteBuf, FoodBlockDataComponent> PACKET_CODEC = FoodBlockEntity.FoodBlockData.PACKET_CODEC
+	public static final StreamCodec<RegistryFriendlyByteBuf, FoodBlockDataComponent> PACKET_CODEC = FoodBlockEntity.FoodBlockData.STREAM_CODEC
 			.map(FoodBlockDataComponent::new, FoodBlockDataComponent::food_block_data);
 	public static final FoodBlockDataComponent DEFAULT = new FoodBlockDataComponent(FoodBlockEntity.FoodBlockData.DEFAULT);
 }