+++ src/main/java/com/github/theredbrain/foodoverhaul/registry/BlockRegistry.java
@@ -10,6 +10,8 @@ import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceKey;
+import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.Item;
@@ -24,12 +26,16 @@ public class BlockRegistry {
 	public static ResourceKey<Block> GENERIC_FOOD_BLOCK_BLOCK_KEY = ResourceKey.create(Registries.BLOCK, FoodOverhaul.identifier("generic_food_block"));
 	public static ResourceKey<Item> GENERIC_FOOD_BLOCK_ITEM_KEY = ResourceKey.create(Registries.ITEM, FoodOverhaul.identifier("generic_food_block"));
 	public static FoodBlockEntity.FoodBlockData GENERIC_FOOD_BLOCK_DATA = new FoodBlockEntity.FoodBlockData(
-			"minecraft:glowing",
-			600,
-			0,
-			false,
-			false,
-			true,
+			List.of(
+					new MobEffectInstance(
+							MobEffects.GLOWING,
+							600,
+							0,
+							false,
+							false,
+							true
+					)
+			),
 			"minecraft:cookie",
 			"minecraft:iron_sword",
 			"",
@@ -37,6 +43,7 @@ public class BlockRegistry {
 			0,
 			false
 	);
+
 	public static final Block GENERIC_FOOD_BLOCK = registerBlockWithFoodBlockData(GENERIC_FOOD_BLOCK_DATA, GENERIC_FOOD_BLOCK_BLOCK_KEY, GENERIC_FOOD_BLOCK_ITEM_KEY, new GenericFoodBlock(BlockBehaviour.Properties.of().setId(GENERIC_FOOD_BLOCK_BLOCK_KEY)), List.of());
 
 	public static ResourceKey<Block> FOOD_DISPLAY_BLOCK_BLOCK_KEY = ResourceKey.create(Registries.BLOCK, FoodOverhaul.identifier("food_display_block"));
@@ -44,7 +51,7 @@ public class BlockRegistry {
 	public static final Block FOOD_DISPLAY_BLOCK = registerBlock(FOOD_DISPLAY_BLOCK_BLOCK_KEY, FOOD_DISPLAY_BLOCK_ITEM_KEY, new FoodDisplayBlock(BlockBehaviour.Properties.of().setId(FOOD_DISPLAY_BLOCK_BLOCK_KEY).strength(0.3F).noOcclusion().isValidSpawn(Blocks::never).isRedstoneConductor(Blocks::never).isSuffocating(Blocks::never).isViewBlocking(Blocks::never)), List.of());
 
 	private static Block registerBlockWithFoodBlockData(FoodBlockEntity.FoodBlockData foodBlockData, ResourceKey<Block> block_key, ResourceKey<Item> item_key, Block block, List<ResourceKey<CreativeModeTab>> itemGroupList) {
-		Registry.register(BuiltInRegistries.ITEM, item_key, new BlockItem(block, new Item.Properties().setId(item_key).component(FoodOverhaul.FOOD_BLOCK_DATA, new FoodBlockDataComponent(foodBlockData)).stacksTo(1)));
+		Registry.register(BuiltInRegistries.ITEM, item_key, new BlockItem(block, new Item.Properties().setId(item_key).component(FoodOverhaulDataComponents.FOOD_BLOCK_DATA, new FoodBlockDataComponent(foodBlockData)).stacksTo(1)));
 		for (ResourceKey<CreativeModeTab> itemGroup : itemGroupList) {
 			CreativeModeTabEvents.modifyOutputEvent(itemGroup).register(content -> content.accept(block));
 		}